Se corrigen bigInt por Long

+++ src/main/GridAndFlex/src/main/java/org/generation/GridAndFlex/controller/PedidosController.java
@@ -5,7 +5,6 @@ import org.generation.GridAndFlex.service.PedidosService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.math.BigInteger;
 import java.util.List;
 
 @RestController
@@ -24,7 +23,7 @@ public class PedidosController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Pedidos> getPedidoById(@PathVariable BigInteger id) {  // ✅ BigInteger
+    public ResponseEntity<Pedidos> getPedidoById(@PathVariable Long id) {
         return pedidosService.getById(id)
                 .map(ResponseEntity::ok)
                 .orElseGet(() -> ResponseEntity.notFound().build());
@@ -37,7 +36,7 @@ public class PedidosController {
 
     @PutMapping("/{id}")
     public ResponseEntity<Pedidos> updatePedido(
-            @PathVariable BigInteger id,
+            @PathVariable Long id,
             @RequestBody Pedidos pedidoData) {
 
         Pedidos pedido = pedidosService.update(id, pedidoData);
@@ -48,7 +47,7 @@ public class PedidosController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deletePedido(@PathVariable BigInteger id) {
+    public ResponseEntity<Void> deletePedido(@PathVariable Long id) {
         boolean deleted = pedidosService.delete(id);
         return deleted ? ResponseEntity.noContent().build() : ResponseEntity.notFound().build();
     }
