+++ src/main/GridAndFlex/src/main/java/org/generation/GridAndFlex/repository/PedidosRepository.java
@@ -3,8 +3,7 @@ package org.generation.GridAndFlex.repository;
 import org.generation.GridAndFlex.model.Pedidos;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.math.BigInteger;
 
-public interface PedidosRepository extends JpaRepository<Pedidos, BigInteger> {
+public interface PedidosRepository extends JpaRepository<Pedidos, Long> {
 
 }
